Move inherited card theme test to material
